style(community-write): move package for CategoryUi
