+++ features/post/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/post/presentation/models/CategoryUi.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.features.post.presentation.communitywrite
+package com.parksupark.soomjae.features.post.presentation.models
 
 import com.parksupark.soomjae.features.post.domain.models.Category
